renaming variables
